[Blazingfast] rename and a bit redesigne
